backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
pymongo==4.3.3
motor==3.1.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
minio==7.1.14
httpx==0.24.0
sqlalchemy>=1.4,<1.5
psycopg2-binary==2.9.6
apache-airflow==2.5.0
pandas==2.0.0
numpy==1.24.2
scikit-learn==1.2.2
joblib==1.2.0
python-dotenv==1.0.0
